requirements.txt
playwright>=1.40.0
aiohttp>=3.8.0
aiofiles>=23.0.0
pandas>=1.5.0
openpyxl>=3.0.0
xlrd>=2.0.0
opencv-python>=4.8.0
numpy>=1.24.0
Pillow>=10.0.0
requests>=2.31.0
cairosvg>=2.7.0
torch>=2.0.0
transformers>=4.30.0
torchvision>=0.15.0 